feat: #WB2-1778, track Library access event
