+++ frontend/src/features/SideBar/Library/Library.tsx
@@ -1,26 +1,11 @@
+import { useLibrary } from "./useLibrary";
 import { ArrowRight } from "@edifice-ui/icons";
-import {
-  usePaths,
-  useOdeTheme,
-  useLibraryUrl,
-  Image,
-  useSession,
-} from "@edifice-ui/react";
+import { Image } from "@edifice-ui/react";
 import { useTranslation } from "react-i18next";
 
 const Library = () => {
+  const { libraryUrl, imageFullURL, handleClick } = useLibrary();
   const { t } = useTranslation();
-  const sessionQuery = useSession();
-  const { theme } = useOdeTheme();
-  const libraryUrl = useLibraryUrl();
-  const [imagePath] = usePaths();
-
-  // #WB2-1689: add end of year Library gif only for FR users
-  const imageFilename =
-    sessionQuery?.data?.currentLanguage === "fr"
-      ? "image-library-year-end.gif"
-      : "image-library.svg";
-  const imageFullURL = `${imagePath}/${theme?.bootstrapVersion}/${imageFilename}`;
 
   return (
     libraryUrl && (
@@ -34,15 +19,15 @@ const Library = () => {
           alt={t("explorer.libray.img.alt")}
         />
         <p className="m-8">{t("explorer.libray.title")}</p>
-        <a
-          href={libraryUrl}
-          target="_blank"
+        <button
+          type="button"
+          onClick={handleClick}
           rel="noreferrer"
-          className="d-inline-flex align-items-center gap-4 btn btn-ghost-primary py-0 p-0 pe-8"
+          className="btn btn-ghost-primary d-inline-flex align-items-center gap-4 p-8"
         >
           <span>{t("explorer.libray.btn")}</span>
           <ArrowRight />
-        </a>
+        </button>
       </div>
     )
   );
